Заменила компонент поиска на Search
